4.1.14 - Fixes for member coins integer limit

+++ src/main/java/net/skycade/kitpvp/commands/CommandSoup.java
@@ -60,7 +60,7 @@ public class CommandSoup extends SkycadeCommand {
         }
 
         KitPvPStats stats = KitPvP.getInstance().getStats(member);
-        int coins = stats.getCoins();
+        long coins = stats.getCoins();
 
         // Applies soup cost reduction upgrade
         int cost = COST;
